# Merge Array with/without Umpire