… CMakeLists.txt
… examples/hasegawa-wakatani-3d/hw.cxx
… manual/sphinx/user_docs/gpu_support.rst