@bendudson 2021-09-16
Now tentatively declaring victory: This version of `Array` compiles and runs on Cori GPUs, and is I think a minimal modification from what is in `master`. 
- (+) Enables pools to be used, so Umpire isn't called repeatedly. If multi-threaded these use a thread-local arena.
- (-) Arrays can be passed to or used inside RAJA CUDA loops